FIR IDE: move ImportableFqNameClassifier to common
