+++ plugins/kotlin/completion/src/org/jetbrains/kotlin/idea/completion/ReferenceVariantsCollector.kt
@@ -5,7 +5,6 @@ package org.jetbrains.kotlin.idea.completion
 import com.intellij.codeInsight.completion.PrefixMatcher
 import org.jetbrains.kotlin.descriptors.*
 import org.jetbrains.kotlin.idea.codeInsight.ReferenceVariantsHelper
-import org.jetbrains.kotlin.idea.core.ImportableFqNameClassifier
 import org.jetbrains.kotlin.idea.core.KotlinIndicesHelper
 import org.jetbrains.kotlin.idea.core.util.externalDescriptors
 import org.jetbrains.kotlin.idea.resolve.ResolutionFacade
